More UI Changes, Make some attributes ints rather than floats

+++ RougeLike/Source/Characters/Attributes/Property.h
@@ -27,20 +27,20 @@ private:
 
 // Shared
 class HealthProperty : public Property<Health> { };
-class MovementSpeedProperty : public Property<float> { };
+class MovementSpeedProperty : public Property<int> { };
 class AttackDamageProperty : public Property<Damage> { };
 
 // Player - Attributes
 class LevelProperty : public Property<Level> { };
-class DefenceProperty : public Property<float> { };
+class DefenceProperty : public Property<int> { };
 class AttackSpeedProperty : public Property<double> { };
 
 // Player - Level up rate
 class HealthUpRate : public Property<Health> { };
-class DefenceUpRate : public Property<float> { };
+class DefenceUpRate : public Property<int> { };
 class AttackDamageUpRate : public Property<Damage> { };
 class AttackSpeedUpRate : public Property<double> { };
-class MovementSpeedUpRate : public Property<float> { };
+class MovementSpeedUpRate : public Property<int> { };
 
 // Player - Movement
 class ForceProperty : public Property<float> { };
